Add files via upload
This is a rudimentary script to calculate the molecular flux for users new to the field of molecular permeability to learn how this could be done.

DISCLAIMER: This script is prone to over-estimating counts across the membrane from unbiased MD. For accurate changes, we suggest using the package Atomix by Eve Biotek, or similar. We recommend testing this carefully before using and do not claim any responsibility for misuse.